requirements.txt
ultralytics==8.0.196
ollama==0.1.7
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
